Bezier model testing implimented.
The testing functions for the bezier curve model have been setup so that
the damage values it returns can be compared to the old linear model.
This will allow further work in both fine tuning the bezier model and
comparing it to the old linear model to be done.

TODO:
- reduce error of bezier model.
- impliment bezier model into ttk plot generation.
- gather more real gun damage data points to allow for more accurate error
  calculation along the whole range

perhaps:
- add a way to save the bezier curves being generated so that they are
  not re-created everytime a damage value is calculated.