Completion of bezier modelling.
The bezier model testing featured in bez_shot_dam.py has been set up
completely in this version. With this, testing and tweaking of the model
has been completed and is showing much greater accuracy compared to real
game data than it did before, and much greater accuracy compared to the
linear model.

The function to generate bezier curves has been modified to offset the
middle control points on the x axis from the middle of the domain. With
this, an optimal offset of 0.15 was found that provides the greatest
accuracy manually. A bisection search would likely produce much better
results.

function bez_shot_dam has also been modified to select out a damage
coefficient value that lies only in the domain being considered. This
was necessary to ensure that the function would not break where
changing the offset caused the call to sympy.solve to return more
than 1 solution.