Optimisations and Fixes
This version comes with some minor optimisation to reduce the time it
takes to generate the plots. Unfortunately the bottleneck is the
generation of the y axis values (the ttk values). Thus, the
optimisation made here was to reduce the number of data points
being generated.

The PP19 has been moved to the SMG category where it belongs.

A decorator to calculate the time a function takes to run has been
written in utility.py.