Refactored ttk, btk, shot_dam Functions.
This version collapses the varients of ttk, btk and shot_dam into single
functions again and giving them the ability to switch between any fall
off model. This will aid in the implementation and testing of any
additional models.

The option to include ads time in the plot_obj_ttk script has been
properly implemented as a command line argument and an argument to the
ttk function itself.

The changes to the other files are minor; remove a line here, make a one
liner there etc.