Update Model Accuracy Testing Script.
The testing script was called 'bez_shot_dam.py'. This has been renamed to
'model_dam_acc.py' on account of this being more representative of the
functionality it now has.

Speaking of which, the script can now easily handle, relatively speaking,
an arbitrary number of weapons and model types.