Modularisation of code. More Tests.
This version predominantly adds greater modularity to the existing test
scripts to futher facilitate the addition of new guns and models to the
datasets.

The creation of the polyfit_realdat.py script has necessitated the
creation of an additional environment as this script uses scipy, a
module not in use in the main scripts.