Massive code cleanup. Removed bezier requirement.
This commit involved a whole lot of cleaning that has not been completed
so that the progress made could be committed. As such, there is
likely some redundant code and outdated documentation lingering around.
TODOs have been left throughout to remind of what needs doing.

The alternative damage models have been removed as the most recently
derived cubic function is not only much faster, but is also more
accurate than the bezier model.

Tests have been added to the project for a number of functions, but
there are more that should be written to ensure that the main script is
behaving itself.

The scripts that test model accuracy have lost some of their
functionality after the cleaning as the functions were disgusting.